+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 4.1.0 - 2026-08-09
+
+### Changed
+
+- Improve commit behavior
+- Move gg commit conventions from gg_git to gg_one_core
+- Move the git and process plumbing to gg_git
+
 ## 4.0.0 - 2026-08-08
 
 ### Changed